+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Updated to use the Rust 2024 edition.
 - Fixed warning about `unnecessary parentheses around closure body`.
+- Use `std::hint::black_box` instead of the one from `criterion`.